Futbol

+++ carles/src/com/company/Util.java
@@ -21,10 +21,10 @@ public class Util {
     System.out.println("---");
   }
 
-  public static void printMatrix(Double[][] matrix) {
+  public static void printMatrix(double[][] matrix) {
     for (int i = 0; i < matrix.length; i++) {
       for (int j = 0; j < matrix[i].length; j++) {
-        System.out.printf("%d ", matrix[j][i].intValue());
+        System.out.printf("%.2f ", matrix[j][i]);
       }
       System.out.println();
     }
